Extract Timestamp component
