+++ app/views/components/pages/header.rb
@@ -13,7 +13,7 @@ module Pages
       render Container.new do |c|
         c.title { title }
         c.description { description } if description
-        render Timestamp.new published_on: published_on, updated_on: updated_on
+        render Pages::Timestamp.new published_on: published_on, updated_on: updated_on
       end
     end
 
@@ -30,43 +30,6 @@ module Pages
 
       def description(&)
         p(class: "description", &)
-      end
-    end
-
-    class Timestamp < ApplicationComponent
-      include Phlex::Rails::Helpers::TimeTag
-
-      attr_accessor :published_on, :updated_on
-      def initialize(published_on:, updated_on:)
-        @published_on = published_on
-        @updated_on = updated_on
-      end
-
-      def view_template
-        if published_on || updated_on
-          span(class: "block") do
-            if published_on && updated_on
-              plain "Published:"
-              whitespace
-            end
-            if published_on
-              em do
-                time_tag published_on, itemprop: "datePublished", class: "dt-published"
-              end
-            end
-            if updated_on
-              if published_on
-                plain " // "
-              end
-              plain "Updated:"
-              whitespace
-
-              em do
-                time_tag updated_on, itemprop: "dateModified", class: "dt-modified"
-              end
-            end
-          end
-        end
       end
     end
   end
